services/langgraph/main.py
import os
import yaml
from langgraph.graph import StateGraph
from typing import Dict, Any

class LangGraphOrchestrator:
    def __init__(self):
        self.graphs = {}
        self.load_graph_definitions()
    
    def load_graph_definitions(self):
        """Load graph definitions from YAML files"""
        graph_dir = "graph_definitions"
        if os.path.exists(graph_dir):
            for file in os.listdir(graph_dir):
                if file.endswith('.yaml'):
                    with open(os.path.join(graph_dir, file), 'r') as f:
                        graph_def = yaml.safe_load(f)
                        self.graphs[file[:-5]] = graph_def
    
    def create_graph(self, graph_name: str) -> StateGraph:
        """Create a LangGraph from definition"""
        if graph_name not in self.graphs:
            raise ValueError(f"Graph {graph_name} not found")
        
        graph_def = self.graphs[graph_name]
        graph = StateGraph(dict)
        
        # Add nodes based on definition
        for node in graph_def.get('nodes', []):
            graph.add_node(node['name'], self._create_node_function(node))
        
        # Add edges
        for edge in graph_def.get('edges', []):
            graph.add_edge(edge['from'], edge['to'])
        
        return graph.compile()
    
    def _create_node_function(self, node_def: Dict[str, Any]):
        """Create node function based on type"""
        node_type = node_def.get('type', 'default')
        
        if node_type == 'vector.query':
            return self._vector_query_node
        elif node_type == 'vector.upsert':
            return self._vector_upsert_node
        else:
            return self._default_node
    
    def _vector_query_node(self, state: Dict[str, Any]) -> Dict[str, Any]:
        # Vector query implementation
        return state
    
    def _vector_upsert_node(self, state: Dict[str, Any]) -> Dict[str, Any]:
        # Vector upsert implementation
        return state
    
    def _default_node(self, state: Dict[str, Any]) -> Dict[str, Any]:
        return state

orchestrator = LangGraphOrchestrator()