services/langgraph/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.40
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
prometheus-client==0.19.0
pyyaml==6.0.1